[move/#120] EventCategoryType 파일 위치 이동
