+++ app/src/main/java/com/bongtu/baekseo/presentation/record/RecordScreen.kt
@@ -35,7 +35,7 @@ import com.bongtu.baekseo.presentation.record.component.AttendTypeTab
 import com.bongtu.baekseo.presentation.record.component.EventCategoryBar
 import com.bongtu.baekseo.presentation.record.component.RecordListContent
 import com.bongtu.baekseo.presentation.record.component.RecordTopBar
-import com.bongtu.baekseo.presentation.record.type.EventCategoryType
+import com.bongtu.baekseo.core.common.type.EventCategoryType
 import kotlinx.collections.immutable.persistentListOf
 
 @Composable
